refactor: back to yarn workspace
