+++ example/package.json
@@ -28,12 +28,6 @@
     "@types/react-dom": "17.0.2",
     "@types/react-helmet": "6.1.0",
     "@types/styled-components": "5.1.8",
-    "typescript": "3.4.5"
-  },
-  "alias": {
-    "gatsby-image": "../node_modules/gatsby-image",
-    "react": "../node_modules/react",
-    "react-dom": "../node_modules/react-dom/profiling",
-    "styled-components": "../node_modules/styled-components"
+    "typescript": "4.2.3"
   }
 }
