+++ gatsby-image-gallery/TSDX.md
