+++ gatsby-image-gallery/package.json
@@ -5,71 +5,60 @@
   "author": "browniebroke",
   "license": "MIT",
   "repository": "browniebroke/gatsby-image-gallery",
+  "homepage": "https://gatsby-image-gallery.netlify.app",
   "main": "dist/index.js",
-  "module": "dist/index.es.js",
-  "jsnext:main": "dist/index.es.js",
+  "typings": "dist/index.d.ts",
+  "module": "dist/gatsby-image-gallery.esm.js",
+  "files": [
+    "dist",
+    "src"
+  ],
   "engines": {
-    "node": ">=10",
-    "npm": ">=5"
-  },
-  "publishConfig": {
-    "access": "public"
+    "node": ">=10"
   },
   "scripts": {
-    "test": "cross-env CI=1 SKIP_PREFLIGHT_CHECK=true react-scripts test --env=jsdom",
-    "test:watch": "SKIP_PREFLIGHT_CHECK=true react-scripts test --env=jsdom",
-    "build": "rollup -c",
-    "start": "rollup -c -w",
-    "prepare": "yarn run build"
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test --passWithNoTests",
+    "lint": "tsdx lint",
+    "prepare": "tsdx build",
+    "size": "size-limit",
+    "analyze": "size-limit --why"
   },
   "peerDependencies": {
     "gatsby-image": ">= 2.0",
-    "prop-types": "^15.5.4",
-    "react": "^16.8 || ^17.0.0",
-    "react-dom": "^16.8 || ^17.0.0",
+    "react": ">=16",
     "styled-components": ">= 5.0"
   },
   "dependencies": {
-    "react-image-lightbox": "^5.1.1",
-    "styled-components": "^5.0.0"
+    "react-image-lightbox": "5.1.1",
+    "styled-components": "5.0.0"
   },
-  "homepage": "https://github.com/browniebroke/gatsby-image-gallery#readme",
   "devDependencies": {
-    "@babel/core": "7.13.8",
-    "@babel/helper-call-delegate": "7.12.13",
-    "@babel/plugin-proposal-class-properties": "7.13.0",
-    "@babel/plugin-proposal-do-expressions": "7.12.13",
-    "@babel/plugin-proposal-export-default-from": "7.12.13",
-    "@babel/plugin-proposal-export-namespace-from": "7.12.13",
-    "@babel/plugin-proposal-function-bind": "7.12.13",
-    "@babel/plugin-proposal-function-sent": "7.12.13",
-    "@babel/plugin-proposal-json-strings": "7.13.8",
-    "@babel/plugin-proposal-logical-assignment-operators": "7.13.8",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "7.13.8",
-    "@babel/plugin-proposal-numeric-separator": "7.12.13",
-    "@babel/plugin-proposal-optional-chaining": "7.13.8",
-    "@babel/plugin-proposal-throw-expressions": "7.12.13",
-    "@babel/plugin-syntax-dynamic-import": "7.8.3",
-    "@babel/plugin-syntax-import-meta": "7.10.4",
-    "@babel/preset-env": "7.13.9",
-    "@babel/preset-react": "7.12.13",
-    "@svgr/rollup": "5.5.0",
-    "cross-env": "7.0.3",
+    "@size-limit/preset-small-lib": "4.10.1",
+    "@types/react": "17.0.3",
+    "@types/react-dom": "17.0.2",
+    "@types/react-test-renderer": "17.0.1",
+    "@types/styled-components": "5.1.8",
+    "babel-jest": "^26.6.3",
     "gatsby-image": "2.11.0",
-    "postcss": "7.0.35",
     "react": "17.0.1",
     "react-dom": "17.0.1",
-    "react-scripts": "4.0.3",
     "react-test-renderer": "17.0.1",
-    "rollup": "2.40.0",
-    "rollup-plugin-babel": "4.4.0",
-    "rollup-plugin-commonjs": "10.1.0",
-    "rollup-plugin-node-resolve": "5.2.0",
-    "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-postcss": "3.1.8",
-    "rollup-plugin-url": "3.0.1"
+    "size-limit": "4.10.1",
+    "tsdx": "0.14.1",
+    "tslib": "2.1.0",
+    "typescript": "4.2.3"
   },
-  "files": [
-    "dist"
+  "size-limit": [
+    {
+      "path": "dist/gatsby-image-gallery.cjs.production.min.js",
+      "limit": "60 KB"
+    },
+    {
+      "path": "dist/gatsby-image-gallery.esm.js",
+      "limit": "60 KB"
+    }
   ]
 }
