+++ gatsby-image-gallery/src/column.tsx
