+++ gatsby-image-gallery/src/img-wrapper.tsx
