+++ gatsby-image-gallery/src/index.tsx
