+++ gatsby-image-gallery/src/row.tsx
