+++ gatsby-image-gallery/test/index.test.tsx
