+++ gatsby-image-gallery/tsconfig.json
